// src/app/model/user-vm.ts
export class UserVM {
  access_token?: string;
  userName?: string;
  userId?: string;
  roleName?: string;
  firstName?: string;
  lastName?: string;
  phoneNumber?: string;
  IsPhoneConfirmed?: boolean;
  email?: string;
  isEmailVerified?: boolean;
  ".issued": string;
  ".expires": string;
  createdBy?: string;
  refresh_token?:string;
}